Borderline Working
Deleted EndingScreen stuff, will just clear stage and then put closing
label there. Created a teams.txt that is based on someone's predictions,
will update when actual schedule is set. BowlPicker screen now has the
ScrollPane populated with most of the information I want, just need to
properly space it out, make it look good, and fix the CheckBox behavior.
Waiting to see if Ian can make an icon for South Alabama. Changed the
window icon, changed the stage size to be wider. Added a header image to
the OpeningScreen, and a place holder for where you will be access the
StandingsTracker screen, hopefully.