BowlPickerController work
Almost exclusively work on the main controller. Checkboxes now function
as intended, and the GUI is probably in its final state. The submit
button is set up but I still need to output the picks to file and email,
and figure out how the emailing works, if it even does. Also mostly
JavaDoc'd BowlPickerController.