GUI work
A lot of progress on the GUI. Fundamental design is now in place, some
mode types supported, but more to come. Biggest problems right now to
tackle are the bracket showing in the TextField, saving/restoring
coefficient inputs, moving formulas to a text file and importing and
exporting them, and supporting the rest of the mode types.