Initial ADDPlayer Push
Currently has support for changing the number of seconds per song and
the number of songs, and then it plays the game, allowing the user to
assign points as needed. Still no support for more than one player,
changing the directory of the music files to be used, keeping the same
song from showing up twice, accessing the metadata of the songs to
display that information to the user, and a closing screen once the game
completes.