Quit round button and preview pane UI changes
Standardized the preview pane element height to 160 regardless of number
of songs. Added a scroll bar when the number of songs is greater than 5.
Added a quit round button to the game-play controller. Need to figure
out why the scroll doesn't work sometimes and also how to finalize the
preview pane prior to showing the results screen (things keep coming in
even after we have left the game-play screen)